refactor: #158 - Reorganiza alguns componentes preexistentes
