+++ client/src/app/(auth)/login/page.tsx
@@ -1,5 +1,5 @@
 import LoginPageComponent from '@/features/login/login';
-import { DotBackground } from '@/shared/components/background';
+import { DotBackground } from '@/shared/components/custom/dot-background';
 
 export default function LoginPage() {
   return (
