--- client/src/shared/components/background.tsx
@@ -1,8 +1,0 @@
-export const DotBackground = () => {
-  return (
-    <>
-      <div className="absolute top-0 z-[-2] h-screen w-screen bg-[size:20px_20px] dark:bg-[#000000] dark:bg-[radial-gradient(#ffffff33_1px,#080808_1px)]"></div>
-      <div className="absolute inset-0 -z-10 h-full w-full bg-white bg-[radial-gradient(#e5e7eb_1px,transparent_1px)] [background-size:16px_16px]"></div>
-    </>
-  );
-};
