+++ client/src/shared/components/custom/checkbox.tsx
@@ -1,28 +1,19 @@
 import * as Label from '@radix-ui/react-label';
 import { Checkbox } from '@/shared/components/ui/checkbox';
 import { cn } from '@/shared/lib/utils';
-import { useId } from 'react';
+import React, { useId } from 'react';
 
-type SimpleCheckboxProps = {
-  id?: string;
+interface SimpleCheckboxProps extends React.ComponentProps<typeof Checkbox> {
   label: string;
-  checked?: boolean;
-  onChange?: (value: boolean) => void;
-  name?: string;
-  className?: string;
   inputClassName?: string;
-  disabled?: boolean;
-};
+}
 
 export function SimpleCheckbox({
   id,
   label,
-  checked,
-  onChange,
-  name,
   className,
   inputClassName,
-  disabled,
+  ...props
 }: SimpleCheckboxProps) {
   const generatedId = useId();
   const checkboxId = id ?? generatedId;
@@ -31,14 +22,11 @@ export function SimpleCheckbox({
     <div className={cn('flex h-[48px] flex-row items-center gap-2', className)}>
       <Checkbox
         id={checkboxId}
-        name={name}
-        checked={checked}
-        onCheckedChange={onChange}
         className={cn(
           'peer data-[state=checked]:bg-brand-500 data-[state=checked]:border-brand-500 rounded-[2px] border-[2px] border-neutral-400',
           inputClassName,
         )}
-        disabled={disabled}
+        {...props}
       />
       <Label.Root
         htmlFor={checkboxId}
