+++ client/src/shared/components/custom/loading-spinner.tsx
@@ -1,21 +1,19 @@
 import { cn } from '@/shared/lib/utils';
 
-export interface ISVGProps extends React.SVGProps<SVGSVGElement> {
+interface LoadingSpinnerProps extends React.SVGProps<SVGSVGElement> {
   size?: number;
-  className?: string;
 }
 
-export const LoadingSpinner = ({
+export function LoadingSpinner({
   size = 24,
   className,
   ...props
-}: ISVGProps) => {
+}: LoadingSpinnerProps) {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
       width={size}
       height={size}
-      {...props}
       viewBox="0 0 24 24"
       fill="none"
       stroke="currentColor"
@@ -23,8 +21,9 @@ export const LoadingSpinner = ({
       strokeLinecap="round"
       strokeLinejoin="round"
       className={cn('animate-spin', className)}
+      {...props}
     >
       <path d="M21 12a9 9 0 1 1-6.219-8.56" />
     </svg>
   );
-};
+}
